Allow user to force db library through LIBS
Until this commit, Xastir would *only* use a DB library if it were
explicitly named one of the things in its loop, and would use the
latest version installed even if it was a mismatch with the headers
specified with --with-bdb-incdir.

This commit precedes that check loop with a test to see if an
appropriate db_create function can be found in the libraries already
in LIBS.

This allows a user to *force* configure to use a version of db of the
user's choice, bypassing the loop.

e.g. on my system, which as both libdb-5.3 and libdb-18.1 installed,
I can *force* Xastir to build against db5 with:

```
../Xastir/configure --with-bdb-incdir=/usr/local/include/db5 --with-bdb-libdir=/usr/local/lib CFLAGS="-O2 -g" CC=gcc12 CXX=g++12 LIBS=-ldb-5.3
```
even though libdb-18.1 and libdb-5.3 are both in /usr/local/lib.

Prior to this commit, specifying the header path for db5 while db18 is
also installed would cause libdb-18.1 to be linked in even though it
doesn't match the headers.

Further refinement of Xastir/Xastir#217 fix