Update INSTALL.md with new info about Berkeley DB
We now support Berkeley DB 18.1 (the most recent) as well as the very
ancient Berkeley DB 5.x.   The install document said that we only
supported version 5, and had examples of how to build with it on a
specific system.

This commit tweaks the INSTALL.md file to be in line with reality.