Remove commented-out block of configure code
A long time ago (prior to 2006), the XASTIR_COMPILER_FLAGS macro
unconditionally inserted "-g -O2" into CFLAGS, with a comment
"everybody likes '-g -O2', right?'"

This broke any attempt to build debug versions of Xastir, and was
actually unnecessary because AC_PROG_CC automatically puts that in if
the user hasn't specified any CFLAGS anyway, but respects the user's
setting of CFLAGS if they have set it.

That bit got disabled in 2006, but in keeping with standard practice of
the day was commented out instead of being removed.

I'm removing it now.