Remove unneeded autoconf macro
Back in 2003, Jack Twilley did a massive autoconf update to streamline
the configuration.

When he did that, he found a macro "XASTIR_SET_SYSTEM" that does
nothing but determine what system one is using and set a variable
called "system".  That variable is then never, ever used anywhere.

Rather than removing this, Jack just put a comment at the top "Is this
really necessary?" and left it there.

Since it does nothing, and hasn't done anything in over 22 years, I'm
removing it.