Check for defective FCC database records
The most recent FCC database has a number of records that have empty
fields for everything except callsign (this appears to happen for
callsigns that have been cancelled).

For example:
EN|1058281|||N7ZCW|L||||||||||||||||||I||||||

Xastir's callsign lookup function screws up when it encounters this
line and declares all callsigns between this one and the next one in
the index to be not found.

The problematic test is this one:

              if ( (temp[0] < line[pos]) ||
                   ( (temp[0] == line[pos]) && (temp[1] < line[pos+1])))
              {

                // "Callsign Search", "Callsign Not Found!"

Here, it's trying to decide if we've passed the desired callsign in
the lookup.   But in this case, line[pos] is "|", which is greater
than the first character in the callsign.

Testing for "|" before doing the actual test fixes the issue.

Closes GH-229