Add calls to make curl follow redirects
Our cURL usage did not respect and follow redirects.  Some of the OSM
tiled geo files we were creating used http: URLs, but the servers that
handle those now redirect to https.  The end result was that users
creating a brand new install of Xastir with no existing map tile cache
would have failed to get their maps downloaded, where older installs
with well-populated caches would never even try to download them and
not notice the problem.

A quick fix was introduced to change some URLs to https and get things
working again, but the *right* fix is to make our cURL usage follow
http/https redirects.

This commit does that.  A single call to `curl_easy_setopt` that adds
the `CURLOPT_FOLLOWLOCATION` option, in two places (dlm.c and
fetch_remote.c) is needed.  With this change, I find that the
OSM_tiled_mapnik.geo map loads properly with an empty cache, whether
the http or https url is used.

I have not been able to check other, non-OSM-tiled maps, but I believe
this to be a correct fix.  Because I haven't had time to test
everything thoroughly, I'm putting this up as a pull request instead
of committing it directly.   I would appreciate it if people who use
online maps other than OSM tiled maps would test it out.

I'll self-merge the pull request if I hear nothing in a week or so or
if it gets approved by other reviewers.

Closes #230