Add some helpful output when curl fails
A user posted (in GH-232) that Xastir could be constructing bad URLs
due to how it formatted lat/lon in WMS requests.  This turned out to
be verified easily by zooming a worldwide WMS server in to just the
right bounding box.

I found it helpful if Xastir actually *said* what URL was being
rejected when curl calls failed, and added one line of debug output to
make it happen, easily seeing that the issue the OP pointed out did in
fact exist, even if it was a little difficult finding a bounding box
that really had the problem.

It was useful enough that I figured I'd put it in the github repo,
too.

GH-232