database.c split, xastir.h cleanup, and add test infrastructure (#234)
This commit covers these main changes:

GNU Autotest infrastructure and a few sample unit tests
Reduce the use of xastir.h
Move much of the GUI code in database.c into a new file
There should be no functional changes (except for the tests).

Along the way my editor decided to mess with the degrees symbols. I believe the original version had the Windows encoding for the degree symbol and in attempting to fix the unwanted change the encoding switched to the UTF-8 version.

Test Infrastructure
-------------------
This is how this whole endeavor started. I am planning some major work in interface.c and I wanted to have tests available to reduce the risk of breakage. Since we are already using Autotools the use of the built in Auotest tools made sense. Tests can be run with a make check.

Some tests for pad_callsign() and extract_speed_course() in db.c are included as samples and to exercise the test infrastructure. Since this started for my planned work in interface.c there is also placeholder test files for some of that work.

xastir.h
-------
To reduce dependecies between source files (and support testing) some work was done to move declarations out of xastir.h. Function definitions that were not needed were removed and definitions for sound.c and db.c were moved.

db.c split
---------
As part of the effort to reduce dependencies functions that directly call X11 were moved to a new db_gui.c. The remaining function declarations in database.hwere moved to db_funcs.h.