Get github actions working for FreeBSD
This is trickier than for other systems, because FreeBSD has to be run
in a VM under Ubuntu.  So there are two layers of actions going on
here.  It was trickier still to get it running with cleanly enumerated
steps (bootstrap, configure, make, make check) because one is running
separate shell commands inside the vm inside a larger step that is
running the vm itself.  But this seems to work now, so I'm unleashing
it.