Replace deprecated uint16 and uint32
uint16 and uint32 are deprecated, pre-C99 types that used to be used
by libtiff.  The C99 standard, which requires including stdint.h, uses
uint16_t and uint32_t, and libtiff itself fixed that usage as far back
as version 4.3.0 (current version is 4.7.1).

This commit brings Xastir in line with that standard and with libtiff.

The map_tif.c file was the only one that used these two types, and
only because it matched what old versions of libtiff did.