Get compressed posit overlays working right
Compressed posits with numeric overlays should have their overlay
characters sent as the letters a through j instead of the digits 0-9, per
the APRS spec.

An attempt was being made to do this in a lot of special cases, mostly
objects, but the code that did it was duplicated over 20 times in
objects.c, and it was done incorrectly.

No attempt was made to do the same thing to regular station posits when
compressed.

This commit removes all of the duplicated attempts to fix the overlay
character being sent, all of which took place right before a call to
compress_posit.  It places the code that does the fixing up into a new
utility function "compress_group", which is then called inside
compress_posit at the time the compressed posit string is constructed.

There was also broken code in db.c that sometimes incorrectly decoded
compressed overlay characters (but only in the very ancient
"NMEA Data" posit format).

Prior to this commit, Xastir would send station posits in compressed
format without the correct digit-to-alphabetic conversion, which could
not be decoded by any compliant APRS client.

Further, prior to this commit, attempts to send compressed object/item
data with numeric overlays would cause Xastir to transmit invalid
characters outside the normal printable ASCII range.

After this commit, station posits with numeric overlays are transmitted with
the correct alphabetic overlay character and are correctly decoded by Xastir.

After this commit, all formats of objects or items that allow the user to
select a symbol transmit correctly in compressed format and can be decoded
properly.  While Xastir had code peppered throughout objects.c that supposedly
handled overlay characters, only regular objects and probability circles
actually allow you to change the symbol through the GUI, and so these are the
only ones I was able to test.  They all worked.

If one wanted to verify that this commit works (since the code
involved is not unit testable in its current state), one needs to
enable both "transmit compressed posits" in the
File->Configure->Station dialog and the "transmit compressed
objects/items" in the File->Configure->Defaults dialog, then set the
station symbol to one in the alternate symbol table, and change the '\'
to a digit.  Upon transmit one should see the correct symbol with overlay
in the map display.

Similarly, to test that all is well with objects, one needs to create
an assortment of object types, all with alternate symbol table symbols
and numeric overlays.

Closes #239