Move DR calculations out of formatting code
Three separate functions in objects.c format the course and speed
components of the transmit string using almost identical code, but one
of them mixes in computation of object position via dead-reckoning in
a way that is irrelevant to the course/speed formatting.

This commit moves that dead reckoning computation up to the part of
the function that extracts the object or item position for formatting,
and consolidates the calls that format latitude and longitude for
transmit.

By doing this, I'll be able to refactor the course/speed formatting
into reusable code so that the same stuff isn't duplicated across
three functions, and be able to unit test that reusable code.

This commit isn't unit testable by itself, as the code it's in depends
on the object or item already being in the database as a DataRow, and is
intricately tied to the whole of Xastir's UI at the moment.  I have
tested it by hand before and after the change by checking that an
object entered with course and speed is properly updated at each
transmit, and an object without course and speed is just left where it
is placed.

c.f. #250