Create object utilities file and tests
I intend to break up the code in objects.c into reusable components
and remove many instances of duplicated code that can be carefully
unit tested.

This is a first step, extracting the code that formats course and
speed into CSE/SPD (or CSD/... or .../SPD or a null string) into a
function, and developing unit tests for that function.

c.f #250