Use object utility function in Create_object_item_tx_string
Make use of format_course_speed in one of the several places that
course/speed strings are formatted.

Tested by creating several objects with and without valid courses and
speeds and assuring that the transmit strings are correct.  The
format_course_speed function itself is under unit testing.

c.f. #250