Consolidate redundant code
Create_object_item_tx_string starts by computing the latitude and
longitude strings in low precision unconditionally, then, in each
separate block of code that is run when we're doing compressed posits,
discards those precomputed strings and recomputes them in
high-precision.

The code is only called once per object processed and only uses the
low-precision strings in the case where compressed posits are not
being sent, and only uses the high-precision when they are.

It is pointless for every single block of code that does compressed
posits (only one of which will be executed for any given object) to
have separate code clobbering the otherwise unused strings.

This commit makes ONE call to the functions that compute the strings,
with the appropriate precision for the value of
transmit_compressed_objects_items that is active.  This simplifies the
many blocks of code that are otherwise nothing more than snprintfs
with different formats and input variables.

Doing this will make refactoring easier.