Extract and unit test code for formatting corridor
Area objects of type 1 or 6 (lines to the left or right) may have a
field of up to three digits, formatted as "{xxx}" representing a
corridor width in miles.

This commit creates a single function that formats a given corridor
width in this manner, and creates unit tests for it.  It then replaces
the code in objects.c's Create_object_item_tx_string function that did
just this with a call to the new function.

The function is fragile, in that it does no sanity checking of its
corridor value.  The data in the DataRow for corridor is a 16-bit
bitfield, so could actually be up to 65535, and
Create_object_item_tx_string only allocates a buffer big enough to
hold three digits plus the brackets, and if passed a value bigger than
three digits the string would be malformed (it would be truncated
after 5 characters and be an invalid corridor).  The only saving grace
is that the code that handles the dialog box that would create these
objects refuses to accept more than three digits.

The code should probably be cleaned up to sanity check the corridor
size and make sure a malformed packet is never created.