Remove unnecessary temporary variables
The code for area objects in Create_object_item_tx_string has now been
reduced to a small number of lines, many of which are function calls
to reusable functions, but most of which remain are two snprintfs that
differ mostly only the format string and two more snprintfs for
compressed versions that also differ from each other only by the
format string.

To prepare for further simplification and consolidation of code, this
commit just removes the unnecessary copying of the "object_group" character
into a temporary that is used for nothing other than to pass to the
snprintf.