Split pthread-specific code out of utils.c
The various "critical section" helper functions in utils.c are
specific to pthreads and so pull in a lot of other code when utils.c
is linked in.   This can complicate unit testing of otherwise simple
functions in util.c or otherwise simple functions that *call*
functions in util.c.

This commit splits all of the mutex-related code into mutex_utils.c
and mutex_utils.h, and makes all the code that *uses* those functions
and structures include mutex_utils.h.

Further work on #257