Split MGRS-related conversions out into their own file
The functions convert_xastir_to_MGRS_str and
convert_xastir_to_MGRS_str_components manipulate the global variable
"coordinate_system" and therefore complicate use of utils.c in unit
testing.

This commit migrates them to their own file and header.

It should be noted that compilers are throwing warnings about
uninitialized variable usage in these functions, and so they need some
work.  I am not touching them yet because this exercise is ONLY about
reducing the number of tendrils of big Xastir include files in
utils.c.

The problematic usages are that the "utmZone" char argument used to be
an internal variable, and there are places where "strlen(utmZone)" is
used where it should be making use of the buffer size passed in.
These should easily be fixable.