Make distance_from_my_station independent of global vars
distance_from_my_station referenced a global variable "english_units"
defined in main.c, forcing utils.c to include main.h in all its glory.

This commit adds an "english_units" parameter to the function and
changes all the calls to it so they pass the variable in instead of
making us read a global.

This further reduces util.c's dependence on main.h, which I really
want it to stop including.

Related to #257 (we're not actually splitting utils.c up here, but
we *are* making it independent)