Refactor probability and corridor functions
The code for format_altitude, format_area_corridor, and
format_probability_ring_data was initially written by cutting and
pasting the code out of Create_object_item_tx_string exactly as is.
Unit tests were created to make sure they produced the expected output
when given appropriate inputs.

But those functions were wonky in that they made use of memcpy,
strcpy, or strcat, which can be fragile, and were full of lines of
code that stuffed a null character into the end of the string
repetitively.

These rewrites simply use snprintf (well, xastir_snprintf, which is
mostly a macro that expands to snprintf on all modern systems) to
write the string in the first place.  snprintf guarantees that the
output string will be terminated unless one passes a zero size, which
we are not doing.

The rewritten functions pass the unit tests.