Extract area object/item formatting code
Simply cut and paste the code that formats objects and
items (compressed and uncompressed) from all the bits and bobs already
formatted out of Create_object_item_tx_string and into a function.

Add unit tests for that function.  There are not enough of them, as
they only check circles of one color, and don't include tests of
objects that can have corridors.

Area objects are not allowed to have course and speed entered from the
dialog box, and if one attempts to add them in the "modify" box (which
does not grey out the course and speed fields) they get ignored.  Area
objects are not allowed to have them, and Xastir enforces that on
input.  Create_object_item_tx_string also zeros out course and speed
if we're doing an area object anyway.

Thus it is not necessary for me to test what happens if you
pass them to this function (I am certain it will blindly insert such
fields even though they are illegal).

This function is not yet used in Create_object_item_tx_string, I want
to make more unit tests first.