Add more area object unit tests
The only thing that hasn't been tested so far for the correct
operation of the new function format_area_object_item_packet when
passed legal arguments is the case where a corridor has been provided.

This commit adds tests of area objects of type 1 (line, angled right,
offset down, with optional corridor), compressed, uncompressed, with
and without altitude.  These tests pass.

Unfortunately, as it is, the function does NO checking that it
actually has legal arguments and depends on the caller to format
things legally.  There's no point in testing that it does the wrong
thing with the wrong inputs, which it certainly does.

If ever this function is going to be beefed up to do error checking,
then is the time to add new unit tests.

For now, since the caller *does* correctly format all the inputs,
we've got proper coverage.