Make use of format_area_object_item_packet
Now, Create_object_item_tx_string calls format_area_object_item_packet
instead of having in-line printfs to do the formatting in the four
special cases (object uncompressed, object compressed, item
uncompressed, item compressed).

All this does is encapsulate the snprintfs into a function, nothing is
really changed but code organization.

All the formatting of the individual elements of the packet is handled
prior to calling this, just as it was prior to the snprintfs before
this change.