Move item name padding to a function
This just moves a block of code down into a function where it can be
unit tested and maybe simplified.

The block of code padded an item callsign with spaces so it is at
least three characters.  This same operation is done in a different
way but to the same effect in Setup_item_data.  Ultimately I'll make
them both call the same function instead of having two blocks of
different code to do the same thing.

Those unit tests are added in this commit as well.

I will likely refactor this function promptly.