Vastly simplify pad_item_name
The only time this function touches the item name is when it's shorter
than three characters, and then all it does is add spaces to make it
three characters.

This is trivially done with snprintf in one line.  And the unit tests
verify that it continues to perform exactly as expected.