Move rng/phg prepending to a new function
If present in the station data record, PHG or RNG data extensions must
be transmitted in the APRS packet by prepending it to the comment.

Xastir doesn't actually allow you to enter a PHG or RNG from the
dialog box (unless you manually code it into the comment yourself), so
it doesn't seem like we should even need to do this, but it is
possible that we received an object from elsewhere that had such an
extension and adopted it, and therefore we still need to keep
transmitting that extension.

At any rate, all it takes to do this is to prepend the power_gain from
the station record to the comment, truncating to 43 characters if
necessary.

All I've done here is move the code to a new function and write some
unit tests.  I will probably refactor the function later.