Add function to format signpost values
format_signpost just takes the signpost string (1-3 chars) and tries
to format it as "{xxx}"

This code cut and pasted out of Create_object_item_tx_string, which
does no error checking to make sure that the value is actually 3
characters, and does the deed in such a way that if it IS too big,
it'd create a malformed packet.  But in this iteration we just do
exactly what has been done forever.

Unit tests in place.

This will be refactored and have error checking inserted in the next
commit.