Add sanity checking to format_area_corridor and format_signpost
The standard requires that area corridors and signpost values be 1-3
characters wide, and formatted as "{xxx}".

Xastir was depending on our inability to *enter* more than three
characters for these values in its dialog box, but the code that
formatted them didn't check, and if the data happened to be
larger (if, for example, someone messed with the dialog box code),
then the text would be truncated to fit the 6 character buffer (with
terminating null) and the result would be a malformed data extension.

This commit adds sanity checking to the functions, which will now
return a null string if the input data is invalid instead of a badly
formatted string.

Format_signpost is also simplified from its original implementation,
and still passes the unit tests.