Use new beam DF object formatting function
Replace actual formatting code in Create_object_item_tx_string for
beam DF objects with call to new, unit-tested function.

Note that as currently written and as originally in place, the code
will scribble "000/000" into the speed_course character buffer if
that buffer didn't have a 7-character string.

This is probably unsafe to maintain in that new function lest one try
to unit test it with a null string literal, but I don't do that in the
test cases.  A simple change can make it safer, and I'll do that
shortly.