Stop clobbering input string in beam df formatting code
format_beam_df_object_item_string would scribble "000/000" into the
speed/course string buffer it is passed if the string didn't have
seven characters, which is just what the Create_object_item_tx_string
code did before I moved it into a new function.

This commit leaves the string it's passed alone, and uses the
"000/000" string literal instead of the pointer to the passed buffer
if the passed string is too short.

This removes the possibility of the function being misused and
resulting in damaged strings or illegal memory accesses if it is
passed something that shouldn't be changed, and leaves all of its
arguments untouched.