Isolate "append comment" block of code in a function
The last thing that Create_object_item_tx_string does before declaring
victory is to append the comment portion of the object's DataRow (if
present) to the packet payload it has assembled, taking care not to
exceed the limits of payload length in the standard.

This commit isolates that code in a function, creates some basic unit
tests for it, and uses it in Create_object_item_tx_string.

I note that the code that was present originally has a comment that it
needs to be modified "when compressed mode is implemented for
objects/items" because those packets have shorter maximum length.

Compressed object/items has been implemented beginning around commit
7e2d5d in 2002, but this comment was put in place before Xastir was
ever in CVS.  So apparently this code does NOT take into account
compressed mode maximum payload size for objects, and may allow
overlong packets.

My unit tests do NOT check this.  ALl they do is check that a comment
that *should* fit in object and item form with uncompressed posit is
not truncated, and that a comment that is far too big *is* truncated.

At some point this should be revisited with test driven development
and careful attention to the spec, which indicates that with
compressed position, the maximum length of an object posit should be
74 characters, and an item should be 58 characters plus however long
the item's name is (that is, a compressed object or item's maximum
length is 6 characters shorter than an uncompressed object or item's).

This function could certainly be written to be more straightforward
than it is, and that should happen *after* writing more tests and
fixing the issue identified above.