Make format_area_color_from_numeric always return valid color string
format_area_color_from_numeric was extracted from code that *presumed*
the color was never over 15, because it was using a 4-bit unsigned
bitfield to store the value and that could never happen.

When I extracted it into a function, I added error checking to guard
against misuse, but if it were misused it would just return a null
string and no information.

Now it prints a warning to stderr and returns a valid color string
anyway (the code for bright red).

Unit tests updated to account for this change.