Clean up object_utils unit test file
For some reason I can't recall, I had been reluctant to include
"object_utils.h" in the file of unit tests for object_utils.c, and
instead simply duplicated the forward declarations of function
templates.

This is unnecessary because "object_utils.h" contains nothing but
those declarations, and so is safe to include in the test harness.

Making it so.