Move CHECKMALLOC and CHECKREALLOC macros to globals.h
Many files had cut/pasted copies of the CHECKMALLOC macro, which does
nothing more than check that a pointer is not null and print an error
message if it is, then cause a fatal error exit.

Consolidating this to one definition in globals.h, making sure that
every file that uses the macro includes the file.

CHECKREALLOC is only used in one file, but I moved it here as well.