Remove stale debugging code from shp_hash.c
When I first wrote shp_hash.c in 2004 I peppered it with debugging
printfs and an include file that accessed private datastructures in
the hashtable setup.  Rather than remove the fprintfs, which are of no
value except to what I happened to be doing 21 years ago, I commented
them out.

They haven't been uncommented since the shapefile hashing system was
debugged and unleashed as default behavior in Xastir in 2006, and
definitely not since the "rtree" system was made the ONLY shapefile
reading method in 2019.

In the interests of clean code, I'm removing those comments.