Remove unused X11 include from lang.c
For as long as Xastir has been in version control, the "lang.c" file
has included "<Xm/XmAll.h>".

This file uses NO X11 data structures or functions, so it is not
necessary to include this header, which brings in *all* of the Motif
headers under the sun.

I am removing this include, because some of the unit testing I am
going to want to start doing on util.c requires that lang.o be linked
in (rather than stubbing out the lang functions).

lang.c is almost completely self-contained so it really should be
something that can be linked in to unit testing with no problem.