Start fleshing out datarow constructor
This commit creates tests for create_object_item_datarow for the very
simplest possible object or item:  having only a name, lat/lon, and
symbol.

We check that the data row is properly populated for this type of
object or item, check that it is or isn't flagged as active as
expected, and finally check that Create_object_item_tx_string formats
the packet correctly from the data given.

Note that because we're actually linking in util.c (so we can get the
functions that convert lat/lon strings to xastir coordinates) we
aren't stubbing out sec_now() anymore, and so we have to clobber the
timestamp produced because the function is using the real sec_now()
before checking the packet against a fixed string.

ONLY very simple objects are correct at the moment.  If one were to
try to pass other data (speed, course, etc.) to the constructor it
would not properly populate the record.