Add speed/course setting to constructor
Add simple tests that an object or item specified with speed and
course gets populated correctly and produces a correct packet.

Add code to constructor to populate the record correctly if speed and
course are specified.

Had to create a fake function to prevent Create_object_item_tx_string
from trying to dead reckon the packet when speed is present.

It is not at all clear from db.c exactly *what* format is legal for
speed and course strings, because the code has so many different
places where it sets that string and they are all over the map as far
as what they will actually add (%06.2f?  %03.0f?  WTF?).   But for our
purposes with objects and items, the dialog box only allows three
characters, so we are going to use %03d for course and %3d for speed,
because at least one place in db.c does it that way, and it works in
the tests.