Add code to process input altitude
The DataRow field for altitude is a string representing the altitude
in meters with two decimal places.

This gets converted back to feet by Create_object_item_tx_string.

Note that Setup_object_data and Setup_item_data do NO such conversion,
because they are constructing the transmit string themselves directly,
and the spec says the altitude in the "/A=xxxxxx" text is supposed to
be in feet.

This commit adds tests of object and item when altitude is specified
in feet.  These tests check that the DataRow gets the correct value of
altitude in meters, and also checks that Create_object_item_tx_string
produces the correct packet string with altitude in feet.

Update code in construct_object_item_data_row to process altitude if
given.  Update comments to reflect my understanding of how the datarow
is supposed to store the altitude.