Add code to support comments in basic object/items
The comment field may be specified in the Object/Item dialog up to 43
characters.  It may be truncated upon transmit if the total packet
length would exceed spec.

This commit adds two tests, one for object and one for item, that add
a comment to the record and test proper function.

"Proper function" is determined by my having created objects with
exactly these parameters in Xastir as it exists today and observing
the packets created.  Those are the gold standards tested in the test
cases.  We check that the full comment string as specified by the user
winds up in the data row, and that the correctly truncated comment
gets put into the packet text by Create_object_item_tx_string when
passed that data row.

Code added to construct_object_item_data_row to make the test pass.

"Destructor" function added to delete the comment (which is
dynamically allocated) before deleting the data row.  Now all tests
call the destructor on the DataRow, even those that could have used
free.