Add support for signposts
A signpost object's DataRow is exactly the same as for a basic object,
except that it may have a single extra bit of signpost data of up to
three characters, and must have a symbol of '\m' (no overlay).

As such, testing of such objects need not be more than to check that
the signpost data is correctly populated and transmitted, as all the
other stuff about regular (not-area) objects is already being put
through that testing.

Xastir incorrectly parses incoming signpost objects that have
altitudes, even though it can correctly create packets that contain
them (it fails to find the signpost text in the comment and doesn't
populate the datarow properly when it receives one, leaving it as an
unprocessed bit of comment), but that's another matter and is already
reported in #259.

This commit tests signpost objects with attached signpost string (the
only thing that distinguishes them from simple objects), with and
without course and speed.  I do not bother checking them with
altitude, again because there's no difference between a DataRow for a
simple object and a signpost in that regard.

It adds the tiny bit of extra processing to the constructor to add the
signpost data to the DataRow.

The arguments in the constructor that are passed this data are chosen
because they're exactly what the dialog box provides (a boolean noting
that it's a signpost, and the extra string data).