Add support for omni-df objects to constructor
Omni DF objects have an "shgd" field that appears as "DFSxxxx/" on
output.  However, the GUI only creates the XXXX part, and the DFSxxxx
part is all that is supposed to get stuffed into the DataRow.

This commit adds tests of the constructor with the df_object and
omni_df flags set and with an shgd string of DFS4133, checking that
the constructor puts exactly that into the data row, and that
Create_object_item_tx_string actually creates the correct transmit
string (i.e. adding the trailing "/").

Add code to constructor to populate the appropriate field of the
DataRow.

Add code to the constructor to set the symbol correctly (the dialog
ONLY permits the symbol of "/\" (DF symbol), so we enforce that).