Add code to handle probability rings
Probability rings are just like regular objects, except the comment
gets a little Xastir-specific string tacked on to the front,
"PminX,PmaxY" where X and Y are radii in miles entered as-is from the
dialog box.

This commit adds code to test that this feature of the constructor
works and produces a DataRow that Create_object_item_tx_string
actually uses to create a correct packet for these objects.

"Correct packet" means "the same packet that I get when I enter
exactly these values into the dialog box in Xastir as it currently is
coded, without using my new constructor."

This completes work on the constructor.

It is now necessary to get Setup_object_data to use this constructor
and Create_object_item_tx_string to make its packet instead of having
nearly-duplicated code from the original version of
Create_object_item_tx_string and Create_item_data.  That will be the
remaining work on this branch.

None of that later work can be unit tested, it'll be laborious work to
test it through the GUI, and the work will be to make sure that we're
reading data from the GUI and constructing exactly the strings that
need to be passed to the constructor in each case.