Create function to pull data from dialog, test it
This commit is a checkpoint.  The new code is compilable and working,
but is not in fact being used yet.  This checkpoint lets me save my
work while I do more extensive manual testing, because this work is
NOT unit-testable (it is all GUI hackage).

We create a new function, Read_object_item_dialog_values, that does
nothing but pull all the data out of the Object/Item
create/modify/delete dialog box and return that data in its arguments.

As it stands in this commit, Setup_object_data still does all its
stuff its original way, but I've added a block of code that:
  - calls Read_object_item_dialog_values using all block-local
  variables as outputs
  - prints out every single thing it has received
  - calls construct_object_item_data_row using all the values it has
  received to make a DataRow.
  - calls Create_object_item_tx_string on this data row and saves
  the resulting transmit string in "experimental line"
  - prints that experimental line

At the end of the function we also print the line that the original
code will produce (and which is actually used for transmit) so I can
verify that both methods produce the same string.

So far, every object I have tried to create has shown the experimental
line to be identical to the original code's line.  We're even doing
dead reckoning of the object correctly.

Once I've tested this some more, the next step is to rip out the old
code and just let the new code do its job.  That will *vastly*
simplify the Setup_object_data code and make addressing the
Setup_item_code issue (#255) trivial (there will be exactly one
character of difference, the flag that says "generate an item packet"
instead of "generate an object packet").